Show CI commands for per-model validators in CI's Gist Post
- Modify validator scripts to output validation commands to a file suffixed "cmd".
- Display the bash command ran to the user.

An alternative is to create the commands directly from the Go command
generator package.

However, I thought that it's easier to let the script that runs the
validator commands output the commands directly. This has two
advantages:
1. The bash script has view of the whole command. In Go we output "$@"
or "$1", which the various `test.sh` files then fill in as validator
script arguments not visible to the Go file.
2. Is is easier to `echo` out the commands in bash than to create the
necessary Go plumbing for `cmd_gen` to be accessed by `post_results`,
which is run as a different process during the CI pipeline.